invoice counts out of common
