+++ packages/invoices/src/InboundShipment/api/operations.graphql
@@ -565,3 +565,14 @@ mutation deleteInboundShipmentLines($input: BatchInboundShipmentInput!) {
 #     }
 #   }
 # }
+
+query invoiceCounts($timezoneOffset: Int) {
+  invoiceCounts(timezoneOffset: $timezoneOffset) {
+    inbound {
+      created {
+        today
+        thisWeek
+      }
+    }
+  }
+}
